+++ LuiData/src/_data_namespace.lua
@@ -260,8 +260,8 @@ local ZoneTable = {}
 --- @class (partial) LuiData
 LuiData = {}
 LuiData.name = "LuiData"
-LuiData.version = 7212
-LuiData.addonVersion = "7.2.1.2"
+LuiData.version = 7213
+LuiData.addonVersion = "7.2.1.3"
 
 --- @class (partial) Data
 LuiData.Data =