+++ LuiExtended/modules/UnitFrames/_CreateCustomFrames.lua
@@ -852,6 +852,59 @@ local function CreateBossFrames()
     end
 end
 
+local SHIELD_BAR_FRAME_BASE_NAMES = { "player", "reticleover", "companion", "SmallGroup", "RaidGroup", "boss", "AvaPlayerTarget", "PetGroup" }
+
+local function ApplyCustomFrameHealthShieldBar(frame, shieldOverlay)
+    local powerBar = frame[COMBAT_MECHANIC_FLAGS_HEALTH]
+    if not powerBar or not powerBar.shield then
+        return
+    end
+
+    powerBar.shield:ClearAnchors()
+    if shieldOverlay then
+        powerBar.shield:SetParent(powerBar.backdrop)
+        if UnitFrames.SV.CustomShieldBarFull then
+            powerBar.shield:SetAnchor(TOPLEFT, powerBar.backdrop, TOPLEFT, 1, 1)
+            powerBar.shield:SetAnchor(BOTTOMRIGHT, powerBar.backdrop, BOTTOMRIGHT, -1, -1)
+        else
+            powerBar.shield:SetAnchor(BOTTOMLEFT, powerBar.backdrop, BOTTOMLEFT, 1, 1)
+            powerBar.shield:SetAnchor(BOTTOMRIGHT, powerBar.backdrop, BOTTOMRIGHT, -1, -1)
+            powerBar.shield:SetHeight(UnitFrames.SV.CustomShieldBarHeight)
+        end
+    else
+        if not powerBar.shieldbackdrop then
+            powerBar.shieldbackdrop = frame.control:CreateControl("$(parent)HealthShieldBackdrop", CT_BACKDROP)
+            powerBar.shieldbackdrop:SetCenterColor(0, 0, 0, 0.4)
+            powerBar.shieldbackdrop:SetEdgeColor(0, 0, 0, 0.6)
+            powerBar.shieldbackdrop:SetEdgeTexture("", 8, 1, 1, 1)
+            powerBar.shieldbackdrop:SetDrawLayer(DL_BACKGROUND)
+            powerBar.shieldbackdrop:SetHidden(true)
+        end
+        powerBar.shield:SetParent(powerBar.shieldbackdrop)
+        powerBar.shield:SetAnchor(TOPLEFT, powerBar.shieldbackdrop, TOPLEFT, 1, 1)
+        powerBar.shield:SetAnchor(BOTTOMRIGHT, powerBar.shieldbackdrop, BOTTOMRIGHT, -1, -1)
+    end
+    powerBar.shield:SetDrawLevel(HEALTH_BAR_FILL_DRAW_LEVEL + 2)
+end
+
+function UnitFrames.CustomFramesApplyShieldBarMode()
+    local baseNameKey = nil
+    local baseName
+    while true do
+        baseNameKey, baseName = next(SHIELD_BAR_FRAME_BASE_NAMES, baseNameKey)
+        if baseNameKey == nil then break end
+        local shieldOverlay = (baseName == "RaidGroup" or baseName == "boss") or not UnitFrames.SV.CustomShieldBarSeparate
+
+        for i = 0, 12 do
+            local unitTag = (i == 0) and baseName or (baseName .. i)
+            local frame = UnitFrames.CustomFrames[unitTag]
+            if frame then
+                ApplyCustomFrameHealthShieldBar(frame, shieldOverlay)
+            end
+        end
+    end
+end
+
 -- Helper to set up common actions for all created frames
 local function SetupCommonFrameActions()
     local function tlwOnMoveStart(self)
@@ -911,9 +964,6 @@ local function SetupCommonFrameActions()
                 LUIE.ApplyPositionLabelFont(unitFrame.tlw.preview.anchorLabel)
             end
         end
-
-        -- Anchor bars to their backdrops
-        local shieldOverlay = (baseName == "RaidGroup" or baseName == "boss") or not UnitFrames.SV.CustomShieldBarSeparate
 
         for i = 0, 12 do
             local unitTag = (i == 0) and baseName or (baseName .. i)
@@ -1021,30 +1071,6 @@ local function SetupCommonFrameActions()
                             end
                         end
 
-                        if powerBar.shield then
-                            powerBar.shield:ClearAnchors()
-                            if shieldOverlay then
-                                if UnitFrames.SV.CustomShieldBarFull then
-                                    powerBar.shield:SetAnchor(TOPLEFT, powerBar.backdrop, TOPLEFT, 1, 1)
-                                    powerBar.shield:SetAnchor(BOTTOMRIGHT, powerBar.backdrop, BOTTOMRIGHT, -1, -1)
-                                else
-                                    powerBar.shield:SetAnchor(BOTTOMLEFT, powerBar.backdrop, BOTTOMLEFT, 1, 1)
-                                    powerBar.shield:SetAnchor(BOTTOMRIGHT, powerBar.backdrop, BOTTOMRIGHT, -1, -1)
-                                    powerBar.shield:SetHeight(UnitFrames.SV.CustomShieldBarHeight)
-                                end
-                            else
-                                powerBar.shieldbackdrop = frame.control:CreateControl("$(parent)ShieldBackdrop", CT_BACKDROP)
-                                powerBar.shieldbackdrop:SetCenterColor(0, 0, 0, 0.4)
-                                powerBar.shieldbackdrop:SetEdgeColor(0, 0, 0, 0.6)
-                                powerBar.shieldbackdrop:SetEdgeTexture("", 8, 1, 1, 1)
-                                powerBar.shieldbackdrop:SetDrawLayer(DL_BACKGROUND)
-                                powerBar.shieldbackdrop:SetHidden(true)
-                                powerBar.shield:SetAnchor(TOPLEFT, powerBar.shieldbackdrop, TOPLEFT, 1, 1)
-                                powerBar.shield:SetAnchor(BOTTOMRIGHT, powerBar.shieldbackdrop, BOTTOMRIGHT, -1, -1)
-                            end
-                            -- Draw shield above no-healing overlay and health fill (HEALTH_BAR_FILL_DRAW_LEVEL = 10)
-                            powerBar.shield:SetDrawLevel(HEALTH_BAR_FILL_DRAW_LEVEL + 2)
-                        end
                     end
                 end
 
@@ -1060,6 +1086,8 @@ local function SetupCommonFrameActions()
             end
         end
     end
+
+    UnitFrames.CustomFramesApplyShieldBarMode()
 end
 
 -- Generic function to setup regen/degen animations